requirements.txt
mysql-connector-python == 8.0.19
pyTelegramBotAPI == 3.7.3